+++ gatsby-config.js
@@ -184,6 +184,23 @@ module.exports = {
         allPageHeaders: ['Referrer-Policy: no-referrer-when-downgrade'],
       },
     },
+
+    {
+      resolve: `gatsby-build-newrelic`,
+      options: {
+        NR_LICENSE_KEY: process.env.NEW_RELIC_LICENSE_KEY,
+        NR_ACCOUNT_ID: process.env.NEW_RELIC_ACCOUNT_ID,
+        SITE_NAME: 'opensource.newrelic.com',
+        collectTraces: true,
+        collectLogs: true,
+        collectMetrics: true,
+        staging: true,
+        customTags: {
+          featureFlag: process.env.FEATURE_FLAG,
+        },
+      },
+    },
+
     // Github API for sponsorship section on homepage
     {
       resolve: `gatsby-source-github-api`,